# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * orDough Bakery API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuItem {
  id: string;
  name: string;
  emoji: string;
  category: string;
  price: number;
  cost: number;
  /** @nullable */
  description?: string | null;
  available: boolean;
  createdAt: string;
}

export interface CreateMenuItemBody {
  name: string;
  emoji: string;
  category: string;
  price: number;
  cost: number;
  /** @nullable */
  description?: string | null;
  available?: boolean;
}

export interface UpdateMenuItemBody {
  name?: string;
  emoji?: string;
  category?: string;
  price?: number;
  cost?: number;
  /** @nullable */
  description?: string | null;
  available?: boolean;
}

export interface Customer {
  id: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  totalOrders: number;
  totalSpent: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreateCustomerBody {
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateCustomerBody {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface OrderItem {
  /** @nullable */
  menuItemId?: string | null;
  name: string;
  emoji: string;
  qty: number;
  price: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  ready: 'ready',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: string;
  /** @nullable */
  customerId?: string | null;
  customerName: string;
  date: string;
  items: OrderItem[];
  discount: number;
  deliveryCharge: number;
  status: OrderStatus;
  /** @nullable */
  notes?: string | null;
  invoiceNumber: string;
  createdAt: string;
}

export type CreateOrderBodyStatus = typeof CreateOrderBodyStatus[keyof typeof CreateOrderBodyStatus];


export const CreateOrderBodyStatus = {
  pending: 'pending',
  ready: 'ready',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface CreateOrderBody {
  /** @nullable */
  customerId?: string | null;
  customerName: string;
  date?: string;
  items: OrderItem[];
  discount?: number;
  deliveryCharge?: number;
  status?: CreateOrderBodyStatus;
  /** @nullable */
  notes?: string | null;
}

export type UpdateOrderBodyStatus = typeof UpdateOrderBodyStatus[keyof typeof UpdateOrderBodyStatus];


export const UpdateOrderBodyStatus = {
  pending: 'pending',
  ready: 'ready',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface UpdateOrderBody {
  /** @nullable */
  customerId?: string | null;
  customerName?: string;
  date?: string;
  items?: OrderItem[];
  discount?: number;
  deliveryCharge?: number;
  status?: UpdateOrderBodyStatus;
  /** @nullable */
  notes?: string | null;
}

export interface Ingredient {
  id: string;
  name: string;
  unit: string;
  stock: number;
  costPerUnit: number;
  minStock: number;
  createdAt: string;
}

export interface CreateIngredientBody {
  name: string;
  unit: string;
  stock: number;
  costPerUnit: number;
  minStock: number;
}

export interface UpdateIngredientBody {
  name?: string;
  unit?: string;
  stock?: number;
  costPerUnit?: number;
  minStock?: number;
}

export interface RecipeIngredient {
  ingredientId: string;
  ingredientName: string;
  quantity: number;
  unit: string;
}

export interface Recipe {
  id: string;
  /** @nullable */
  menuItemId?: string | null;
  menuItemName: string;
  ingredients: RecipeIngredient[];
  /** @nullable */
  instructions?: string | null;
  servings: number;
  createdAt: string;
}

export interface CreateRecipeBody {
  /** @nullable */
  menuItemId?: string | null;
  menuItemName: string;
  ingredients: RecipeIngredient[];
  /** @nullable */
  instructions?: string | null;
  servings: number;
}

export interface UpdateRecipeBody {
  /** @nullable */
  menuItemId?: string | null;
  menuItemName?: string;
  ingredients?: RecipeIngredient[];
  /** @nullable */
  instructions?: string | null;
  servings?: number;
}

export interface Review {
  id: string;
  customerName: string;
  rating: number;
  comment: string;
  /** @nullable */
  product?: string | null;
  approved: boolean;
  createdAt: string;
}

export interface CreateReviewBody {
  customerName: string;
  rating: number;
  comment: string;
  /** @nullable */
  product?: string | null;
  approved?: boolean;
}

export interface UpdateReviewBody {
  customerName?: string;
  rating?: number;
  comment?: string;
  /** @nullable */
  product?: string | null;
  approved?: boolean;
}

export interface Settings {
  id: string;
  bakeryName: string;
  tagline: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  upiId?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  currency: string;
  taxRate: number;
  updatedAt: string;
}

export interface UpdateSettingsBody {
  bakeryName?: string;
  tagline?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  upiId?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  currency?: string;
  taxRate?: number;
}

export interface DashboardSummary {
  totalOrders: number;
  pendingOrders: number;
  todayRevenue: number;
  totalRevenue: number;
  totalCustomers: number;
  /** @nullable */
  topSellingItem?: string | null;
  lowStockCount: number;
  recentOrders: Order[];
}

export type ProfitReportByCategoryItem = {
  category: string;
  revenue: number;
  cost: number;
  profit: number;
};

export type ProfitReportByMonthItem = {
  month: string;
  revenue: number;
  orders: number;
};

export interface ProfitReport {
  totalRevenue: number;
  totalCost: number;
  grossProfit: number;
  profitMargin: number;
  byCategory: ProfitReportByCategoryItem[];
  byMonth: ProfitReportByMonthItem[];
}

export type ListCustomersParams = {
search?: string;
};

export type ListOrdersParams = {
status?: ListOrdersStatus;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending: 'pending',
  ready: 'ready',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;