# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: orDough Bakery API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # --- CUSTOMERS ---
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /customers/{id}/orders:
    get:
      operationId: getCustomerOrders
      tags: [customers]
      summary: Get orders for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer's orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  # --- ORDERS ---
...
  /ingredients:
    get:
      operationId: listIngredients
      tags: [ingredients]
      summary: List all ingredients
      responses:
        "200":
          description: List of ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ingredient"
    post:
      operationId: createIngredient
      tags: [ingredients]
      summary: Create an ingredient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIngredientBody"
      responses:
        "201":
          description: Created ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
  /ingredients/{id}:
    patch:
      operationId: updateIngredient
      tags: [ingredients]
      summary: Update an ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIngredientBody"
      responses:
        "200":
          description: Updated ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
        "404":
          description: Not found
    delete:
      operationId: deleteIngredient
      tags: [ingredients]
      summary: Delete an ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # --- RECIPES ---
...
components:
  schemas:
...
    MenuItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        emoji:
          type: string
        category:
          type: string
        price:
          type: number
        cost:
          type: number
        description:
          type: ["string", "null"]
        available:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - emoji
        - category
        - price
        - cost
        - available
        - createdAt
    CreateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        emoji:
          type: string
        category:
          type: string
        price:
          type: number
        cost:
          type: number
        description:
          type: ["string", "null"]
        available:
          type: boolean
      required:
        - name
        - emoji
        - category
        - price
        - cost
    UpdateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        emoji:
          type: string
        category:
          type: string
        price:
          type: number
        cost:
          type: number
        description:
          type: ["string", "null"]
        available:
          type: boolean
    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        totalOrders:
          type: integer
        totalSpent:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - totalOrders
        - totalSpent
        - createdAt
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
    UpdateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    OrderItem:
      type: object
      properties:
        menuItemId:
          type: ["string", "null"]
        name:
          type: string
        emoji:
          type: string
        qty:
          type: integer
        price:
          type: number
      required:
        - name
        - emoji
        - qty
        - price
    Order:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: ["string", "null"]
        customerName:
          type: string
        date:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        discount:
          type: number
        deliveryCharge:
          type: number
        status:
          type: string
          enum: [pending, ready, delivered, cancelled]
        notes:
          type: ["string", "null"]
        invoiceNumber:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - date
        - items
        - discount
        - deliveryCharge
        - status
        - invoiceNumber
        - createdAt
...
    Ingredient:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        unit:
          type: string
        stock:
          type: number
        costPerUnit:
          type: number
        minStock:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - unit
        - stock
        - costPerUnit
        - minStock
        - createdAt
    CreateIngredientBody:
      type: object
      properties:
        name:
          type: string
        unit:
          type: string
        stock:
          type: number
        costPerUnit:
          type: number
        minStock:
          type: number
      required:
        - name
        - unit
        - stock
        - costPerUnit
        - minStock
    UpdateIngredientBody:
      type: object
      properties:
        name:
          type: string
        unit:
          type: string
        stock:
          type: number
        costPerUnit:
          type: number
        minStock:
          type: number
...