# lib/api-zod/src/generated/types/createMenuItemBody.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * orDough Bakery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMenuItemBody {
  name: string;
  emoji: string;
  category: string;
  price: number;
  cost: number;
  /** @nullable */
  description?: string | null;
  available?: boolean;
}