# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * orDough Bakery API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  /** @nullable */
  customerId?: string | null;
  customerName: string;
  date: Date;
  items: OrderItem[];
  discount: number;
  deliveryCharge: number;
  status: OrderStatus;
  /** @nullable */
  notes?: string | null;
  invoiceNumber: string;
  createdAt: Date;
}