# lib/api-zod/src/generated/types/updateIngredientBody.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * orDough Bakery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateIngredientBody {
  name?: string;
  unit?: string;
  stock?: number;
  costPerUnit?: number;
  minStock?: number;
}